fix: null pointer crash in WM_CREATE; add MinGW runtime DLL auto-copy
- Fix CameraPreviewApp unique_ptr release before use causing crash (0xC000041D)
- Add POST_BUILD step to copy libstdc++-6.dll, libgcc_s_seh-1.dll, libwinpthread-1.dll to build/
- Various source file updates across camera, imaging, ui, storage modules

+++ src/camera/CameraTelemetryFormatter.cpp
@@ -26,7 +26,7 @@ std::wstring CameraTelemetryFormatter::FormatFrameTelemetry(
     int device_index,
     const CameraOpenInfo& open_info,
     double fps,
-    unsigned long timestamp,
+    uint32_t timestamp,
     UILanguage lang)
 {
     std::wostringstream fps_str;
