+++ src/domain/ImageFrame.h
@@ -1,13 +1,14 @@
 #pragma once
 
+#include <cstdint>
 #include <vector>
 
 struct ImageFrame {
     int width = 0;
     int height = 0;
     int stride = 0;
-    unsigned long timestamp = 0;
-    unsigned long long sequence = 0;
+    uint32_t timestamp = 0;
+    uint64_t sequence = 0;
     std::vector<unsigned char> bgr;
 
     bool IsValid() const
